feat: implement can_redeem endpoint
The response of this endpoint determines whether the enroll button is
shown, and contains business logic to decide which access policy should
be used (i.e. the "policy resolver").  It also should produce or
otherwise passthrough helper URLs to do things like check on redemption
status, create a new redemption, and redirect to a courseware view.

ENT-6854