fix: refactor can_redeem to make one call per subsidy to fetch all learner txs
ENT-7221

To find existing redemptions in can_redeem, we now make a single call to list transactions for the learner,
the result of which we map out in a dictionary variable in the view, by content and policy identifiers,
and then store in the request cache.  Caching per request allows us to reuse these results
later if we need to determine the amount spent by the learner in a policy is below the policy's
learner-level spend cap.