feat: model fields to have correct dependency direction.
Updates the relationship between the SAP and Assignments modules so that
SAP -> AssignmentConfiguration, not the other way around.
We achieve this by introducing a new model, AssignmentConfiguration, which
replaces the AssignmentPolicy model.
The SubsidyAccessPolicy model now links, with a 1-1 field, to the
AssignmentConfiguration model (side note: the name "AssignmentConfiguration")
feels much more natural in the context of defining a 1-1, and more accurately
captures the intention of the model).
The FK link from Assignments to AssignmentPolicy is removed in favor of a nullable FK
linking to the AssignmentConfiguration model.
This all works as a single PR because there's no business logic that currently
uses the existing AssignmentPolicy model, and zero records pertaining to Assignments
exist in our non-local systems. ENT-7532