feat: populate assignment.lms_user_id on core.User.save()
During the post_save hook for core.User, lookup any existing assignments
for the user and populate the lms_user_id into the assignments.  This
should happen whenever the user logs in, so it theoretically covers the
case where a learner is assigned content before authing with
enterprise-access.

The case where a learner is assigned content AFTER authing with
enterprise-access would be covered by ENT-7874.

ENT-7875