feat: subsidy access policy description field allowed to be blank
Before this change, editing learner credit plans from the LC
provisioning tool in frontend-app-support-tools had a weird UX where
blanking the description field of a budget/policy would silently
re-populate the field with a calculated value.  This allows the frontend
to just blank the field and not subvert expectations of the enterprise
customer support person, and helps kill some frontend code.

ENT-7922