feat: reminder failures no longer fail assignments
This is a stop-the-bleeding approach to solving the problem where
reminder failures cause the assignment to become inadvertently
non-redeemable. This commit results in the remind and remind-all
endpoints no longer progressing the assignment lifecycle state to
`errored` when the reminder fails for any reason.

Before this PR, clicking remind would cause admins to see the following
table cells change:

* Remidner failure:
  - Status column: "Waiting for learner" -> "Failed: Reminder"
  - Recent action column: "Assigned: <date>" -> NO CHANGE
* Reminder success:
  - Status column: "Waiting for learner" -> NO CHANGE
  - Recent action column: "Assigned: <date>" -> "Reminded: <date>"

but now they will see:

* Remidner failure:
  - Status column: "Waiting for learner" -> NO CHANGE
  - Recent action column: "Assigned: <date>" -> NO CHANGE
* Reminder success:
  - Status column: "Waiting for learner" -> NO CHANGE
  - Recent action column: "Assigned: <date>" -> "Reminded: <date>"

This is obviously not ideal because the failure case causes no visual
change in the admin UI whatsoever. Absence of a POSITIVE change in the
recent action column implies a NEGATIVE reminder outcome, so reminder
failures are *indirectly* inferrable. However, this is likely not the
preferred long-term UX because reminder failures are not *directly*
inferrable.

ENT-8362