fix: uniquify retired email addresses
Prior to this change, the `LearnerContentAssignment.clear_pii()` method set a retired email address
to the same static value.  This means that if two different learners (by email)
in the same assignment configuration (i.e. budget) have the same course assigned,
and then those records are expired, the second record to be expired will fail on the unique
constraint violation of the model.
This change adds a random hex-string to each retired email address value to help resolve this issue.
ENT-8574