fix: allocation identifies learners by either email or lms_user_id
The assignment *expiration* logic modifies the email field of the assignment (but not the lms_user_id).
Prior to this change, when we look for existing assignments during allocation, we checked only by email.
With this change, we now also check for existing, re-allocatable, assignments by lms_user_id.

This commit also fixes a bug where the "*_at" datetime fields weren't
having a null value persisted on them during reallocation.

ENT-9122