feat: can_redeem now checks for enrollment deadline
We want the `can_redeem` query to return false for a given (user, content key)
if the enrollment deadline for the content key is in the past. This is necessary because
some courses have runs with an advertised run for which the deadline is in the future,
but also have available recent runs where the upgrade deadline is in the past.

ENT-7472