fix: [can-redeem] bogus content should not throw 500 response
Previous behavior before we required a non-null list_price was to return
200 response with reason = "content_not_in_catalog" if the content key
was totally fake.  Now that we are no longer tolerant of NOT including
the price in the response, we must return some sort of non-200 response.

Before this PR, the can-redeem endpoint would return 500.  This fixes
the response to return 422 Unprocessable to at least indicate that the
requester did something wrong.

The immediate value in returning 422 vs. 500 is not that it would
benefit our frontends, but rather it provides better feedback to
API-based enterprise customers.

ENT-9660