feat: optional catalog input during provisioning
The caller of the subscription provisioning API is not required
to specify the enterprise catalog record desired. If none is provided,
we infer the catalog title based on a simple format, and infer
the catalog query id based on the requested subscription plan product_id,
which field *is* required. ENT-10519