feat: Minimal CheckoutIntent serializer
ENT-10700 | CheckoutIntent state needs to be surfaced to the frontend
in order to give real-time feedback to the "go to dashboard" stateful button,
conditionally perform a forced redirect from other pages to the success page,
and support a hard refresh of the Success page.
This change adds a minimal CI serializer to the checkout context BFF response.