feat: allow provisioning of multiple subs plans with same opp line item
This still ratains the restriction against multiple plans with the same
opp_line_item and same product_id. This change essentially results in
[opp_line_item, product_id] being "unique together", but on the
client-side instead of server-side.

This change is meant support the use case of provisioning a trial plan
alongside a paid plan with the same opp_line_item.

ENT-11009